move files to separate article folder
