+++ frontend-app/welcome-app/article/article-overview.js
@@ -1,6 +1,6 @@
 import Article from './article.js';
-import { parseArticleOverview } from './helpers/article-overview-parser.js';
-import { articleTypes } from './article/article-types.js';
+import { parseArticleOverview } from '../helpers/article-overview-parser.js';
+import { articleTypes } from './article-types.js';
 
 class ArticleOverview extends Article {
